requirements.txt
einops==0.4.1
numpy==1.19.5
opencv-python==3.4.10.37
Pillow==8.3.2
PyYAML==5.3.1
scikit-learn==0.24.2
scipy==1.5.4
tabulate==0.8.10
tensorboardX==1.8
torch==1.5.0
torchvision==0.6.0+cu90
tensorboardX
tabulate
torch==1.11.0